api/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pyairtable>=2.1.0
python-dotenv>=1.0.0
